@@ -3,17 +3,17 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@fontsource/poppins": "^4.5.0",
-		"@reduxjs/toolkit": "^1.7.1",
-		"date-fns": "^2.28.0",
+		"@fontsource/poppins": "^5.1.0",
+		"@reduxjs/toolkit": "^2.3.0",
+		"date-fns": "^4.1.0",
 		"framer-motion": "^5.5.1",
-		"node-sass": "^7.0.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-hot-toast": "^2.2.0",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
+		"react-hot-toast": "^2.4.1",
 		"react-icons": "^4.3.1",
-		"react-redux": "^7.2.6",
-		"react-scripts": "5.0.0",
+		"react-redux": "^9.1.2",
+		"react-scripts": "^5.0.1",
+		"sass": "^1.79.5",
 		"uuid": "^8.3.2"
 	},
 	"scripts": {
